refactor(audio-reader): extract term-card from the term popover
Pull the translation fetch + card content (term header, reading/pos,
definition states, add-card form) out of term-popover/index.vue into a
presentation-agnostic term-card.vue. The popover now just anchors the
card in a ui-popover surface. This lets a second presentation (a mobile
sheet) reuse the same card. Also i18n the divider's 'Definition' label.
